src: Use UTF-16 instead of UCS-2 for SMS
Although UTF-16 is not mentioned in the 3gpp standards as a valid SMS
coding, most smartphones use it instead of plain UCS-2 when UCS-2 is
the encoding specified in the SMS header. This allows the phone to send
characters out of the basic multiligual plane, like emojis. As these
characters are codified using codes forbidden for UCS-2, all valid
UCS-2 code points can still be sent.