gril: data_call parse error fix/code clean-up
Fixed a parse error when a reply to a REQUEST_SETUP_DATA_CALL was
received with a non-zero status.  When this happens the remaining
settings aren't always populated, thus the code was always logging
a parse error with the returned type ( IP protocol ).  As part of
this fix, the grilreply version of the parsing code was removed,
and now this reply is handled by grilunsol which has similar code
to handle the UNSOL_DATA_CALL_LIST_CHANGED message.  Also modified
the data/voice registration reply parsing to no longer translate radio
technology values to ofono's access/bearer technology constants, this
is now handled by the upper rilmodem atoms.