src: Implement functions needed by MTK driver
Due to the way the current MTK device plugin handles creation of
atoms, some modem state transitions cannot be handled directly by
the core. This change adds a new function that can be used by a
device plugin to prevent the modem core code from setting the modem
state to PRE-SIM when a SIM lock condition is detected.