mtk: Support the 3G caps switch case
Add support in case we change the default slot with 3G capabilities.
Needed changes are:

1. rild sockets do not always control the same physical slot: it can
change depending on the slot that owns the 3G capabilities at a given
time. /dev/socket/rild always controls the 3G slot, while
/dev/socket/rild2 controls the other one. As we want to keep the
association between physical slots 0/1 and object paths /ril_0/1, now
we check the 3G capabilities so we know the physical slot that we
control with a given socket, and we associate the right modem path
with that socket. Having just one slot is also considered so things
work also in that case.

2. A function to reset the modem. This function should be invoked when
we have changed the slot with 3G caps.