mtk: Fix slow socket disconnection in FM
When unsetting flight mode, rild usually closes the sockets, but this
does not necessarily happens always. In some cases this disconnection
is notified to ofono a bit too late, which leaves ofono in a bad state.
This change adds a timeout so in case we actually expect the sockets to
be closed we slow down the sequence of ofono actions lo leave time for
the notification to happen.