gprs: Resolve MMS proxy/MMSC host name
Resolve, in case the hostname is provided instead of the IP address,
MMS proxy / MMSC name using the DNS servers provided by the context.
We need the IP so we can set-up a route for them. Also, use the MMSC
for the route in case we do not have a MMS proxy. This solves LP bug
1417976.