stemodem: Create socket when used
Create netlink socket when the driver is used. Previously was being
created always, which created a conflict when using the netlink library
in the gprs atom. That happened because in both cases the socket ended
up being created in the same nl port (being the port number equal to the
process ID).