plugins: address upower plugin upstream comments
The primary change simplifies the plugin so that it's always
listening for UPower PropertiesChanged signals vs. only doing
so if one or more HFP atoms has been recognized.  The trade-off
is slightly additional overhead due to context switches caused
by the incoming DBus signals ( every 8-10s ) vs. a simpler solution,
which also gets rid of a race condition between the HF and the
plugin.