ril: Create radio setting when onlining
Create/remove the radio setting atom when the modem is onlined/offlined.
We do not follow in this case the usual oFono life cycle, as we need
this atom to exist when we are online (otherwise some of the requests it
uses fail) but also when there is no SIM (so we can enter low power mode
for the modem even in that case).