rilmodem: radio-settings: enable switch slot caps
Enable radio settings atom so now we can switch the capabilities of the
slots in multi-sim devices. This will be mostly transparent to the DBus
clients: the AvailableTechnologies property will show all possible
technologies for a slot, even those that require a caps switch to
happen. When one of those is selected, the capability switch is
performed. As a side effect, the selected technology for the other slot
will change to a different technology if its currect tech is one of
those transferred in the caps switch.

We also set now here the initial attach APN, as now the slot with LTE
capabilities might change dynamically. If we switch to LTE the slot, we
set the IA APN for it too.