rilmodem: radio-settings: implement legacy cap switch
In older versions of RIL (and some newer one), capability switch can be
done simply by setting the other slot's prefered network to GSM. This
commit automates the preference change in other slots to make sure that
the current preference setting will succeed.

To control whether the device (with newer RIL) uses the legacy
capability switch or not, a new environment variable "OFONO_RIL_LEGACY_
CAP_SWITCH" is introduced. When this environment variable is set to non-
zero value, the legacy capability switch will be used.