gril: follow struct size change in data call list
There are changes in RIL's data call list response in version 9 (or 10)
and 11 of the response. This cause a problem when there are more than 1
entry in the list.

This commit checks the version field to find out if it needs to read the
extra field(s). This fixes the problem where devices with newer Android
can randomly become unable to send or receive an MMS message.

Note that there's conflicting information between ril.h [1] and RIL.java
[2] about the availability of the pcscf field. Following ril.h seems to
cause a problem with some Android 4.4 devices. Thus, this version of
commit is updated to follow RIL.java.

[1] https://github.com/LineageOS/android_hardware_ril/blob/cm-14.1/include/telephony/ril.h#L489
[2] https://github.com/LineageOS/android_frameworks_opt_telephony/blob/cm-14.1/src/java/com/android/internal/telephony/RIL.java#L4024