emulator: don't return error after calling cb
Returning error from a function with callback implies that the callback
will not be called. Commit 018272a6dd5c7d1ac2227018614d3473aad4df88
("emulator: fail when SCO connection setup fails") broke this assumption
by returning error from ofono_handsfree_card_connect_sco after the
callback is called with success. This created a double-free situation
in hfp_ag_bluez5.c where it frees its "cbd" both in the callback and
when the call itself fails.

This commit make the code always return success in this situation, and
instead return the success or error via the callback, which seems to be
the code's original intention. I'm not sure if calling callback before
returning from the function is OK, but if the original code works
correctly in non-error case, this should work correctly too.